refactor(api): Improve SeasonalTitle API based on review feedback
- Consolidate seasonalTitle validation schema into skill.ts to eliminate
  duplicate requiredId helper
- Parallelize DB queries in /current endpoint using Promise.all for
  better performance
- Add createdAt and updatedAt fields to /current endpoint response
- Fix comment to accurately describe sort order (minSpEarned desc)
- Update tests to verify new response fields
