+++ src/app/api/skills/seasonal-titles/current/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server'
 import { prisma } from '@/lib/prisma'
-import { seasonalTitlesQuerySchema } from '@/lib/validations/seasonalTitle'
+import { seasonalTitlesQuerySchema } from '@/lib/validations/skill'
 import {
   formatZodError,
   formatNotFoundError,
@@ -35,27 +35,26 @@ export async function GET(request: NextRequest) {
     // 直近N日のdayKeyを取得
     const recentDayKeys = getRecentDayKeys(category.rankWindowDays)
 
-    // 直近N日のSP合計を計算
-    const categoryResults = await prisma.dailyCategoryResult.findMany({
-      where: {
-        categoryId: result.data.categoryId,
-        dayKey: { in: recentDayKeys },
-      },
-      select: {
-        spEarned: true,
-      },
-    })
+    // 直近N日のSP合計と称号一覧を並列で取得
+    const [categoryResults, titles] = await Promise.all([
+      prisma.dailyCategoryResult.findMany({
+        where: {
+          categoryId: result.data.categoryId,
+          dayKey: { in: recentDayKeys },
+        },
+        select: { spEarned: true },
+      }),
+      // 称号一覧を取得（minSpEarned降順で、最も高い称号から確認）
+      prisma.seasonalTitle.findMany({
+        where: { categoryId: result.data.categoryId },
+        orderBy: [{ minSpEarned: 'desc' }, { order: 'desc' }],
+      }),
+    ])
 
     const totalSpEarned = categoryResults.reduce(
       (sum, r) => sum + r.spEarned,
       0
     )
-
-    // 称号一覧を取得（order降順で、minSpEarnedが高い順に確認）
-    const titles = await prisma.seasonalTitle.findMany({
-      where: { categoryId: result.data.categoryId },
-      orderBy: [{ minSpEarned: 'desc' }, { order: 'desc' }],
-    })
 
     // 現在のSP合計以下で最も高い称号を見つける
     const currentTitle =
@@ -69,6 +68,8 @@ export async function GET(request: NextRequest) {
             label: currentTitle.label,
             minSpEarned: currentTitle.minSpEarned,
             order: currentTitle.order,
+            createdAt: currentTitle.createdAt,
+            updatedAt: currentTitle.updatedAt,
           }
         : null,
       totalSpEarned,
