+++ src/lib/validations/skill.ts
@@ -18,6 +18,11 @@ export const skillNodesQuerySchema = z.object({
   treeId: requiredId('ツリーID'),
 })
 
+export const seasonalTitlesQuerySchema = z.object({
+  categoryId: requiredId('カテゴリID'),
+})
+
 export type SkillTreesQueryInput = z.infer<typeof skillTreesQuerySchema>
 export type SkillTreeIdParamInput = z.infer<typeof skillTreeIdParamSchema>
 export type SkillNodesQueryInput = z.infer<typeof skillNodesQuerySchema>
+export type SeasonalTitlesQueryInput = z.infer<typeof seasonalTitlesQuerySchema>
